pyproject.toml
[project]
name = "rss-digest"
version = "0.1.0"
description = "Fetch RSS feeds, deduplicate entries, and deliver summaries via Signal, Slack, or email"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["rss", "atom", "digest", "signal", "slack", "news"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
dependencies = [
    "feedparser>=6.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
rss-digest = "rss_digest.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rss_digest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
